Create temperature monitoring
This data is assimilated by the satellite and ground station together and we can see that temperature is available 2 meters height and also spatial resolution is higher approximate 10 Km. The result of this analysis is used for temporal analysis or time series analysis to know anomaly changes or the percent of changes.